2022/2022-04-mr-salesman/requirements.txt
colorama==0.4.4
pydash==5.1.0
termcolor~=1.1.0
ACO-Pants